requirements.txt
numpy>=1.21.0
spatialmath-python>=1.1.0 